Added userscript for controlling the playback of HTML5 videos on any website.
This commit introduces a userscript that allows users to control the playback of HTML5 videos on any website. The script includes features such as decreasing and increasing video speed, as well as changing the video saturation. Installation instructions are provided, as well as usage guidelines for controlling the video speed. Users are also encouraged to submit pull requests for suggestions or improvements.